Add constructor and state to NewToot
NewToot is going to be where we keep the state for our tiny little form.
We want to initialize the value of `inputValue` as soon as the Component
is mounted into the page, so we'll use the constructor to accomplish it.

Because we want *this component*, the one we're working in right now, to
have the state, we're going to set the state parameter on `this` by
using `this.state =`.

`this` can be a tricky concept in JavaScript, so we'll try and keep it
simple by addressing it when we need to use it here, and not get too
deep in the weeds just yet.