Add updateInput function to NewToot
We're going to need a way to update the value we just put into the state
of NewToot. React gives us a helpful function called `setState` but
rather than writing it directly into the component we're passing it to..
we'll make it a parameter on the class. This way it isn't being defined
inside the render function, and everything stays a little bit more
separated, and a little easier to read.

This also allows us the opportunity to use this function in another
place for another reason. It doesn't look like we'll need that ability
for now, but who knows? Making it separate like this gives us the
readability increase, so we'll put the re-use superpower in our back
pocket for now. Maybe we'll be happy its there later.