Pass updateInput into NewTootInput as prop
We're going to need that function down in the NewTootInput if we want to
change the state in its parent component. Remember that data can only
flow in one direction, that's a theme or goal of the React project. When
you only have one place where new data can come from, and you keep that
direction clean, then you're narrowing the chance of bugs, and you're
also making them easier to find because you have less places to look.

Props can be data, and they can also be functions. That's what we're
taking advantage of here. Notice the `this` again. Because it's a
function defined on our class component, we need to specify that
location when we pass it down.

It can also act as a little reminder to us that the `updateInput`
function belongs to NewToot, which is the place that is holding onto the
state. Our updating functions need to originate there, and get loaned
out as props to other components that might need to use them.