Pass down inputValue state as prop to NewTootInput
We need all of our interactions between our components, and all the data
that they know about to come into React. So even though the input field
can be typed into just fine, we want to make sure that as we update the
state, that is exactly what will be reflected in the input element. So
we're going to send a message along to the child component with the
current value from NewToot's state.

By passing it down as a prop, we're also letting React know that
NewTootInput is going to care when `this.state.inputValue` changes. In a
way it creates a link between that bit of state and the child component.
If that state value changes, React will look and see if there is a
resulting difference or change in how it should display NewTootInput to
the user. If there is, it will update that section of the UI tree using
the new value in state.

That might sound like there's a bit of a circular link, NewTootInput
triggers a change to the state with that prop function, which changes
the prop value that's coming down to it. Well... that's true! It's okay
here, that's exactly what we want.

It does make you think though, doesn't it. What if every time the child
component re-renders, it updated the state again? It would go on and on
forever! Well eventually the browser would run out of memory and crash,
but this is an easy way to make an infinite loop. So if something is
updating a parent's state, and then doing it again everytime it
re-renders that's a recipe for disaster.