A little housekeeping
I moved the props on the input component each onto their own line. A
little because I think it looks nicer that way, and also because when we
add our other props in a second, it would look like a lot more has
changed because of how git/github displays changes as entire lines. So
we're cleaning house before making major changes.

Notice I also put `props` into the argument parens on NewTootInput's
definition. Before our previous changes, there weren't any props coming
into this component so it wasn't necessary, but now there are. By naming
the props object React is going to give us in those parens, it'll give
us access to it when we need it.