Add the new props to the <input>
Now we're putting those props to work.

onChange is a prop that sets an event handler. This means that every
time a "change" event happens, that function will be run, and React will
pass in an event object as it's argument. Take a look at where we
defined that function and you'll see we're already prepared for it.

Bring your attention to the onChange prop now, if it isn't already, and
notice that there aren't a set of parens on this function name. `()`
tells the JavaScript runtime to execute the function as soon as it sees
it, meaning that every time the NewTootInput gets put into the page,
that function would be run! Well we only want it to be passed in as a
value to execute later, so we'll leave off the parens, and that function
will sit there and wait just like any other variable.

Personally, this is something I really love about JavaScript that I miss
when I'm writing in a language like Ruby that doesn't really accommodate
that kind of behavior. JS is referred to as having "first order
functions," meaning this exact thing. Functions can be set to variables
and passed around to be executed later, and not every time the
function's name is written. Opens up a lot of cool possibilities....

The `value` prop here is pretty straightforward. Does what it says on
the tin. It puts the value of our variable, our prop, `tootValue` into
the <input> element, completing the loop of data staying entirely within
React's control. A form field that does that is often referred to as a
"controlled component," and that's how you want your React forms to
work. Remember we step on React's toes and lose some opinionated
benefits if we start changing data or UI outside of React's view.

Really personifying React here.. aren't I?